+++ ta4j-core/src/test/java/org/ta4j/core/indicators/averages/TripleEMAIndicatorTest.java
@@ -21,63 +21,50 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.indicators;
+package org.ta4j.core.indicators.averages;
 
-import static junit.framework.TestCase.assertEquals;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.AbstractIndicatorTest;
+import org.ta4j.core.indicators.averages.TripleEMAIndicator;
 import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
 import org.ta4j.core.mocks.MockBarSeriesBuilder;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.num.NumFactory;
 
-public class ZLEMAIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
+public class TripleEMAIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
 
-    private BarSeries data;
+    private ClosePriceIndicator closePrice;
 
-    public ZLEMAIndicatorTest(NumFactory numFactory) {
+    public TripleEMAIndicatorTest(NumFactory numFactory) {
         super(numFactory);
     }
 
     @Before
     public void setUp() {
-        data = new MockBarSeriesBuilder().withNumFactory(numFactory)
-                .withData(10, 15, 20, 18, 17, 18, 15, 12, 10, 8, 5, 2)
+        var data = new MockBarSeriesBuilder().withNumFactory(numFactory)
+                .withData(0.73, 0.72, 0.86, 0.72, 0.62, 0.76, 0.84, 0.69, 0.65, 0.71, 0.53, 0.73, 0.77, 0.67, 0.68)
                 .build();
+        closePrice = new ClosePriceIndicator(data);
     }
 
     @Test
-    public void ZLEMAUsingBarCount10UsingClosePrice() {
-        var zlema = new ZLEMAIndicator(new ClosePriceIndicator(data), 10);
+    public void tripleEMAUsingBarCount5UsingClosePrice() {
+        var tripleEma = new TripleEMAIndicator(closePrice, 5);
 
-        assertNumEquals(11.9091, zlema.getValue(9));
-        assertNumEquals(8.8347, zlema.getValue(10));
-        assertNumEquals(5.7739, zlema.getValue(11));
-    }
+        assertNumEquals(0.73, tripleEma.getValue(0));
+        assertNumEquals(0.7229, tripleEma.getValue(1));
+        assertNumEquals(0.8185, tripleEma.getValue(2));
 
-    @Test
-    public void ZLEMAFirstValueShouldBeEqualsToFirstDataValue() {
-        var zlema = new ZLEMAIndicator(new ClosePriceIndicator(data), 10);
-        assertNumEquals(10, zlema.getValue(0));
-    }
+        assertNumEquals(0.8027, tripleEma.getValue(6));
+        assertNumEquals(0.7328, tripleEma.getValue(7));
+        assertNumEquals(0.6725, tripleEma.getValue(8));
 
-    @Test
-    public void valuesLessThanBarCountMustBeEqualsToSMAValues() {
-        var zlema = new ZLEMAIndicator(new ClosePriceIndicator(data), 10);
-        var sma = new SMAIndicator(new ClosePriceIndicator(data), 10);
-
-        for (int i = 0; i < 9; i++) {
-            assertEquals(sma.getValue(i), zlema.getValue(i));
-        }
-    }
-
-    @Test
-    public void smallBarCount() {
-        var zlema = new ZLEMAIndicator(new ClosePriceIndicator(data), 1);
-        assertNumEquals(10, zlema.getValue(0));
+        assertNumEquals(0.7386, tripleEma.getValue(12));
+        assertNumEquals(0.6994, tripleEma.getValue(13));
+        assertNumEquals(0.6876, tripleEma.getValue(14));
     }
 }